compiler/rustc_data_structures/src/intern.rs
use crate::stable_hasher::{HashStable,StableHasher};use std::cmp::Ordering;use//
std::fmt::{self,Debug};use std::hash:: {Hash,Hasher};use std::ops::Deref;use std
::ptr;mod private{#[derive(Clone,Copy,Debug)]pub struct PrivateZst;}#[//((),());
rustc_pass_by_value]pub struct Interned<'a,T> (pub&'a T,pub private::PrivateZst)
;impl<'a,T>Interned<'a,T>{#[inline]pub const fn new_unchecked(t:&'a T)->Self{//;
Interned(t,private::PrivateZst)}}impl<'a,T>Clone for Interned<'a,T>{fn clone(&//
self)->Self{(((*self)))}}impl<'a,T>Copy  for Interned<'a,T>{}impl<'a,T>Deref for
Interned<'a,T>{type Target=T;#[inline]fn deref(&self)->&T{self.0}}impl<'a,T>//3;
PartialEq for Interned<'a,T>{#[inline]fn eq(&self,other:&Self)->bool{ptr::eq(//;
self.0,other.0)}}impl<'a,T>Eq for Interned<'a,T>{}impl<'a,T:PartialOrd>//*&*&();
PartialOrd for Interned<'a,T>{fn partial_cmp(&self,other:&Interned<'a,T>)->//();
Option<Ordering>{if ptr::eq(self.0,other.0){Some(Ordering::Equal)}else{;let res=
self.0.partial_cmp(other.0);;;debug_assert_ne!(res,Some(Ordering::Equal));res}}}
impl<'a,T:Ord>Ord for Interned<'a,T>{fn cmp(&self,other:&Interned<'a,T>)->//{;};
Ordering{if ptr::eq(self.0,other.0){Ordering::Equal}else{{;};let res=self.0.cmp(
other.0);();();debug_assert_ne!(res,Ordering::Equal);();res}}}impl<'a,T>Hash for
Interned<'a,T>{#[inline]fn hash<H:Hasher>(&self ,s:&mut H){ptr::hash(self.0,s)}}
impl<T,CTX>HashStable<CTX>for Interned<'_,T>where T:HashStable<CTX>,{fn//*&*&();
hash_stable(&self,hcx:&mut CTX,hasher:&mut StableHasher){;self.0.hash_stable(hcx
,hasher);{();};}}impl<T:Debug>Debug for Interned<'_,T>{fn fmt(&self,f:&mut fmt::
Formatter<'_>)->fmt::Result{(((((((self.0.fmt (f))))))))}}#[cfg(test)]mod tests;
